chore: replace helpers with traits
